requirements.txt
PyYAML
PyGithub
google-cloud-aiplatform
google-auth
GitPython
secure-smtplib
grpcio